django/requirements.txt
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.7.2
asttokens==2.4.1
cachetools==5.3.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.2.1
conf==0.4.1
contourpy==1.2.1
cryptography==41.0.3
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==5.0.2
Django==4.2.8
django-allauth==0.59.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.14.0
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.50.0
google-api-core==2.19.0
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.4
importlib-metadata==7.0.1
importlib_resources==6.4.0
ipykernel==6.28.0
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
Markdown==3.6
matplotlib==3.8.4
matplotlib-inline==0.1.6
mysqlclient==2.2.4
nest-asyncio==1.5.8
numpy==1.26.4
oauthlib==3.2.2
openai==1.30.1
packaging==24.0
pandas==2.2.1
parso==0.8.3
pillow==10.3.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.7
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pygame==2.5.2
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3
pywin32==306
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
soundfile==0.12.1
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
wcwidth==0.2.13
zipp==3.18.1
